% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.r
\name{calibrate}
\alias{calibrate}
\title{A function used in the class Statistics for Laboratory Scientists (140.615/616).}
\usage{
calibrate(x, y, newy)
}
\arguments{
\item{x}{Locations of the design points for the calibration line.}

\item{y}{Measurements of the observed outcomes for the calibration line.}

\item{newy}{New measurement(s) for which the calibration is needed.}
}
\value{
Estimate of the x value for the new y.
}
\description{
This function carries out the linear regression based calibration introduced in class.
}
\examples{
quinine <- c(0,0,0, 12,12,12, 24,24,24, 36,36,36)
fluor <- c(100.45, 98.92,101.33, 133.19,127.33,126.78, 152.72,157.43,160.81, 188.73,191.96,183.70)
calibrate(quinine, fluor, 143.70)
newy <- c(148.56, 149.36, 150.29)
calibrate(quinine, fluor, newy)
}
\keyword{Calibration,}
\keyword{linear}
\keyword{regression.}
