% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer.r
\docType{data}
\name{cancer}
\alias{cancer}
\title{Cancer and survival time data}
\format{
A data frame with 64 rows and 2 columns. The response is numeric, the cancer type is a factor.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
cancer
}
\description{
Patients with advanced cancers of the stomach, bronchus, colon, ovary or breast were treated with ascorbate. The purpose of the study was to determine if patient survival differed with respect to the organ affected by the cancer.
}
\examples{
str(cancer)
summary(cancer)
}
\keyword{datasets}
