% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlorine.r
\docType{data}
\name{chlorine}
\alias{chlorine}
\title{Chlorine data}
\format{
A data frame with 44 rows and 2 columns, the amount of chlorine as a function of time.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
chlorine
}
\description{
Chlorine decay over time.
}
\examples{
chlorine
summary(chlorine)
plot(chlorine$time,chlorine$chlorine)
}
\keyword{datasets}
