% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.bonf.r
\name{ci.bonf}
\alias{ci.bonf}
\title{A function used in the class Statistics for Laboratory Scientists (140.615/616).}
\usage{
ci.bonf(response, group, alpha = 0.05)
}
\arguments{
\item{response}{A numeric outcome.}

\item{group}{A factor specifying the treatment assignment.}

\item{alpha}{The significance level.}
}
\description{
This function calculates the Bonferroni corrected confidence intervals for all pairwise group comparisons, based on a pooled estimate of the within-group standard deviation from a one-way ANOVA.
}
\examples{
summary(sugar)
ci.bonf(sugar$rsp, sugar$trt)
}
\keyword{Bonferroni.}
\keyword{Multiple}
\keyword{comparisons,}
