% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper.r
\docType{data}
\name{copper}
\alias{copper}
\title{Copper data}
\format{
A data frame with 24 rows and 2 columns, the concentration of copper (concentration) and the corresponding measurements of percent transmittance (response).
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
copper
}
\description{
The copper data concern an analysis of copper using flame atomic absorption spectroscopy. There are two columns. the concentration of copper (in ppm) in a set of standards, and corresponding measurements of percent transmittance (via the spectroscopy procedure).
}
\examples{
copper
str(copper)
plot(copper$concentration,copper$response)
}
\keyword{datasets}
