% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.anc1.r
\docType{data}
\name{dat.anc1}
\alias{dat.anc1}
\title{ANCOVA example}
\format{
A data frame with 36 rows and 3 columns. The response (rsp) is numeric, the treatment (trt) is a factor, and age is numeric.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
dat.anc1
}
\description{
In this simulated data set we have three different treatments, and we are interested whether treatment has an effect on a response. In this example, the response might also depend on age.
}
\examples{
str(dat.anc1)
summary(dat.anc1)
tapply(dat.anc1$rsp,dat.anc1$trt,mean)
}
\keyword{datasets}
