% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.bp1.r
\docType{data}
\name{dat.bp1}
\alias{dat.bp1}
\title{Blood pressure data}
\format{
A data frame with 80 rows and 3 columns. The response (bp) is numeric, the predictors (water and diet) are factors.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
dat.bp1
}
\description{
In this simulated data blood pressure is the outcome and two factors (two levels each) are predictors: water (plan and salted) and diet (normal and high fat).
}
\examples{
str(dat.bp1)
summary(dat.bp1)
tapply(dat.bp1$bp,list(dat.bp1$water,dat.bp1$diet),mean)
}
\keyword{datasets}
