% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.bp2.r
\docType{data}
\name{dat.bp2}
\alias{dat.bp2}
\title{Blood pressure data}
\format{
A data frame with 80 rows and 3 columns. The response (bp) is numeric, the predictors (water and diet) are factors.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
dat.bp2
}
\description{
In this simulated data blood pressure is the outcome and two factors (two levels each) are predictors: water (plan and salted) and diet (normal and high fat).
}
\examples{
str(dat.bp2)
summary(dat.bp2)
tapply(dat.bp2$bp,list(dat.bp2$water,dat.bp2$diet),mean)
}
\keyword{datasets}
