% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.ex3.r
\docType{data}
\name{dat.ex3}
\alias{dat.ex3}
\title{Longitudinal data}
\format{
A data frame with 120 rows and 4 columns. The response (y) and the main predictor (x) are numeric, the other predictors (id and group) are factors.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
dat.ex3
}
\description{
Simulated data to demonstrate the ultility of the lmer R function for linear mixed effects in a longitudinal data analysis.
}
\examples{
str(dat.ex3)
summary(dat.ex3)
table(dat.ex3$group,dat.ex3$id)
}
\keyword{datasets}
