% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.plot.r
\name{dot.plot}
\alias{dot.plot}
\title{A function used in the class Statistics for Laboratory Scientists (140.615/616).}
\usage{
dot.plot(x, y, includeCI = TRUE, labels = c("A", "B"), xlim)
}
\arguments{
\item{x}{Measurements for the first group.}

\item{y}{Measurements for the second group.}

\item{includeCI}{Whether or not to include a 95\% confidence interval. The default is TRUE.}

\item{labels}{Plot labels for the two groups. Default is "A" and "B".}

\item{xlim}{Optional argument for the limits of the x axis.}
}
\value{
The two-group comparison dotplot.
}
\description{
This function generates the two group comparison dot plot used in class.
}
\examples{
x <- c(102.5,106.6,99.8,106.5,103.7,105.5,98.2,104.1,85.6,105.5,114.0,112.2)
y <- c(93.7,90.9,100.4,92.0,100.2,104.6,95.4,96.6,99.2)
dot.plot(x,y)
}
\keyword{Dotplots,}
\keyword{confidence}
\keyword{intervals.}
\keyword{means,}
