% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flies.r
\docType{data}
\name{flies}
\alias{flies}
\title{Flies data}
\format{
A data frame with 192 rows and 3 columns. The response (response) is numeric, the predictors (far and strain) are factors.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
flies
}
\description{
Data from an experiment measuring size of flies among eight strains. Three jars per strain, eight flies per jar.
}
\examples{
str(flies)
summary(flies)
tapply(flies$response,list(flies$jar,flies$strain),mean)
}
\keyword{datasets}
