% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heme2.r
\docType{data}
\name{heme2}
\alias{heme2}
\title{Heme data}
\format{
A data frame with 24 rows and 3 columns, the optical density (od) and the concentration of H2O2 (conc) are numeric, the species of heme (heme) is recorded as a factor.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
heme2
}
\description{
For two species of heme the optical density was recorded as a function of H2O2 concentration.
}
\examples{
heme2
str(heme2)
summary(heme2)
plot(heme2$conc, heme2$od, xlab="concentration", ylab="optical density", col=rep(c("red","blue"),rep(12,2)))
}
\keyword{datasets}
