% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meioses.r
\docType{data}
\name{meioses}
\alias{meioses}
\title{Meioses data}
\format{
A data frame with 92 rows and 3 columns. The data have three columns: family, female, male. These last two give the total number of crossovers on the 22 autosomes in each of the male and female meioses. All columns are integers.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
mosq
}
\description{
Data from an experiment measuring the total number of crossovers on the 22 autosomes in each of the male and female meioses.
}
\examples{
meioses
str(meioses)
summary(meioses)
tapply(meioses$female,meioses$family,mean)
tapply(meioses$male,meioses$family,mean)
}
\keyword{datasets}
