% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosq.r
\docType{data}
\name{mosq}
\alias{mosq}
\title{Mosquito data}
\format{
A data frame with 24 rows and 3 columns. The response (length) is numeric, the predictors (cage and individual) are factors.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
mosq
}
\description{
Data from an experiment measuring length of mosquito wings (two measurements per mosquito). Four mosquitos per cage. Three cages total.
}
\examples{
mosq
str(mosq)
summary(mosq)
tapply(mosq$length,list(mosq$individual,mosq$cage),mean)
}
\keyword{datasets}
