% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse.r
\docType{data}
\name{mouse}
\alias{mouse}
\title{Hibernating pocket mouse data}
\format{
A data frame with 24 rows and 3 columns. The response (rsp) is numeric, the predictors (food and temperature) are factors.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
mouse
}
\description{
The data were obtained in a study on energy utilization (in kcal/g) of the pocket mouse during hibernation at different temperatures.
}
\examples{
mouse
str(mouse)
summary(mouse)
tapply(mouse$rsp,list(mouse$food,mouse$temp),mean)
}
\keyword{datasets}
