% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.anova.r
\name{nested.anova}
\alias{nested.anova}
\title{A function used in the class Statistics for Laboratory Scientists (140.615/616).}
\usage{
nested.anova(aov.output)
}
\arguments{
\item{aov.out}{The output from the aov() function for nested ANOVA.}
}
\value{
The correct nested ANOVA table.
}
\description{
This function corrects the nested ANOVA table obtained from the aov() R function. It calculates the correct F statistic and p-value for the nested effect.
}
\examples{
data(mosq)
mosq.aov <- aov(length ~ cage / individual, data=mosq)
summary(mosq.aov)

# The above is wrong. Below is the correct table.

nested.anova(mosq.aov)
}
\keyword{ANOVA,}
\keyword{Nested}
\keyword{effects.}
\keyword{random}
