% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newman.keuls.r
\name{newman.keuls}
\alias{newman.keuls}
\title{A function used in the class Statistics for Laboratory Scientists (140.615/616).}
\usage{
newman.keuls(response, group, alpha = 0.05)
}
\arguments{
\item{response}{A numeric outcome.}

\item{group}{A factor specifying the treatment assignment.}

\item{alpha}{The significance level.}
}
\description{
This function plots the results from the Newman Keuls procedure for a one-way ANOVA.
}
\examples{
summary(sugar)
newman.keuls(sugar$rsp, sugar$trt)
}
\keyword{Multiple}
\keyword{comparisons.}
