% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npk.fac.r
\docType{data}
\name{npk.fac}
\alias{npk.fac}
\title{NPK data}
\format{
A data frame with 24 rows and 4 columns. Recorded are the presence or absence (1/0) of nitrogen (N), phosphate (P) and potassium (K) and the yield.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
npk.fac
}
\description{
Data from a classical N, P, K (nitrogen, phosphate, potassium) factorial experiment on the growth of peas, with replicates.
}
\examples{
npk.fac
str(npk.fac)
summary(npk.fac)
}
\keyword{datasets}
