% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pear.r
\docType{data}
\name{pear}
\alias{pear}
\title{Pearson's data of fathers' and daughters' heights}
\format{
A data frame with 1,376 rows and 2 columns.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
pear
}
\description{
The heights of fathers and their daughters, reported by Karl Pearson around 1903.
}
\examples{
str(pear)
summary(pear)
}
\keyword{datasets}
