% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowers.r
\docType{data}
\name{rowers}
\alias{rowers}
\title{Rowers data}
\format{
A data frame with 9 rows and 2 columns, the height (given in cm) and weight (given in kg) of the crew.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
rowers
}
\description{
Data of the crew of the Germany Eight rowing boat.
}
\examples{
rowers
summary(rowers)
plot(rowers$height,rowers$weight)
}
\keyword{datasets}
