% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sediment.r
\docType{data}
\name{sediment}
\alias{sediment}
\title{Mud snail data}
\format{
A data frame with 18 rows and 2 columns, the amount of pellets ingested (pellets) in the time allowed (time in minutes).
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
sediment
}
\description{
Sediment ingestion by the mud snail, Hyrobia minuta.
}
\examples{
sediment
summary(sediment)
plot(sediment$time,sediment$pellets)
}
\keyword{datasets}
