% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.r
\docType{data}
\name{simdat}
\alias{simdat}
\title{A simulated example for a two-way mixed effects ANOVA with replicates}
\format{
A data frame with 100 rows and 3 columns. The response is numeric, the predictors treat and block are factors.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
simdat
}
\description{
Five replicates for each of four treatments (A-D; fixed effects) recorded on each of 5 separate days (blocks B1-B5; random effects).
}
\examples{
str(simdat)
summary(simdat)
tapply(simdat$response,list(simdat$treat,simdat$block),mean)
}
\keyword{datasets}
