% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span.r
\docType{data}
\name{span}
\alias{span}
\title{Pearson's fathers' height and span data}
\format{
A data frame with 1,050 rows and 2 columns.
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
span
}
\description{
The heights (stature) and spans of fathers, reported by Karl Pearson around 1903.
}
\examples{
head(span)
str(span)
summary(span)
}
\keyword{datasets}
