% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiders.r
\docType{data}
\name{spiders}
\alias{spiders}
\title{Spider mites data}
\format{
A data frame with 7 rows and 3 columns. Recorded are the doses of DDT (dose), the number of dead spider mites in each batch after conclusion of the experiment (n.dead), and the number of spider mites per batch (n).
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
spiders
}
\description{
Data from an experiment exposing batches of 25 spider mites to different doses of DDT.
}
\examples{
spiders
str(spiders)
summary(spiders)
spiders$n.dead/spiders$n
}
\keyword{datasets}
