% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticks.r
\docType{data}
\name{ticks}
\alias{ticks}
\title{Ticks data}
\format{
A data frame with 8 rows and 4 columns. Recorded are the tick sex (Tsex, fe for female and ma for male), the leg the gland substance came from (Leg, fo for fore leg, hi for hind leg), the deer sex (Dsex, fe for female and ma for male), number of ticks going to the treated tube (T) and the number of ticks going to the untreated tube (U).
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
ticks
}
\description{
Data from the tick experiment. Place ticks on clay island surrounded by water, with two capillary tubes: one treated with deer-gland-substance, and one untreated.
}
\examples{
ticks
str(ticks)
summary(ticks)
ticks$T/(ticks$T+ticks$U)
}
\keyword{datasets}
