% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worms.r
\docType{data}
\name{worms}
\alias{worms}
\title{Tobacco budworm data}
\format{
A data frame with 12 rows and 4 columns. Recorded are the doses (dose), the number of dead worms in each batch after conclusion of the experiment (n.dead), the number of worms per batch (n) and the sex of the worms (sex).
}
\source{
\href{https://biostat.jhsph.edu/~iruczins/teaching/140.615/}{Class page}
}
\usage{
worms
}
\description{
Data from an experiment involving the tobacco budworm, Heliothis virescens. Batches of 20 male and 20 female worms were given a 3-day dose of pyrethroid trans-cypermethrin The number dead in each batch was recorded.
}
\examples{
worms
worms$n.dead/worms$n
}
\keyword{datasets}
